+++ test/Math/Combinatorics/permutation.jl
@@ -1,7 +1,7 @@
 function test_permutation_permutations()
-    @test collect(permutations(Int[], Val{:unique}())) == Array{Int64,1}[[]]
+    @test collect(permutations(Int[], Val{:unique}())) == Array{Int,1}[[]]
     @test length(permutations(Int[], Val{:unique}())) == 1
-    @test eltype(permutations(Int[], Val{:unique}())) == Array{Int64,1}
+    @test eltype(permutations(Int[], Val{:unique}())) == Array{Int,1}
 
     @test collect(permutations(["a", "b", "c"], Val{:unique}())) == Array[
         ["a", "b", "c"],
@@ -16,9 +16,9 @@ function test_permutation_permutations()
 end
 
 function test_permutation_permutations_with_repetition()
-    @test collect(permutations(Int[], Val{:repeated}())) == Array{Int64,1}[[]]
+    @test collect(permutations(Int[], Val{:repeated}())) == Array{Int,1}[[]]
     @test length(permutations(Int[], Val{:repeated}())) == 1
-    @test eltype(permutations(Int[], Val{:repeated}())) == Array{Int64,1}
+    @test eltype(permutations(Int[], Val{:repeated}())) == Array{Int,1}
 
     @test collect(permutations(["a", "b", "c"], Val{:repeated}())) == Array[
         ["a", "a", "a"],
@@ -58,13 +58,13 @@ function test_permutation_permutations_unknown_mode()
 end
 
 function test_permutation_kpermutations()
-    @test collect(permutations(Int[], 0, Val{:unique}())) == Array{Int64,1}[[]]
+    @test collect(permutations(Int[], 0, Val{:unique}())) == Array{Int,1}[[]]
     @test length(permutations(Int[], 0, Val{:unique}())) == 1
-    @test eltype(permutations(Int[], 0, Val{:unique}())) == Array{Int64,1}
+    @test eltype(permutations(Int[], 0, Val{:unique}())) == Array{Int,1}
 
-    @test collect(permutations(Int[], 1, Val{:unique}())) == Array{Int64,1}[]
+    @test collect(permutations(Int[], 1, Val{:unique}())) == Array{Int,1}[]
     @test length(permutations(Int[], 1, Val{:unique}())) == 0
-    @test eltype(permutations(Int[], 1, Val{:unique}())) == Array{Int64,1}
+    @test eltype(permutations(Int[], 1, Val{:unique}())) == Array{Int,1}
 
     @test collect(permutations(["a", "b", "c", "d"], 2, Val{:unique}())) == Array[
         ["a", "b"],
@@ -117,13 +117,13 @@ function test_permutation_kpermutations()
 end
 
 function test_permutation_kpermutations_with_repetition()
-    @test collect(permutations(Int[], 0, Val{:repeated}())) == Array{Int64,1}[[]]
+    @test collect(permutations(Int[], 0, Val{:repeated}())) == Array{Int,1}[[]]
     @test length(permutations(Int[], 0, Val{:repeated}())) == 1
-    @test eltype(permutations(Int[], 0, Val{:repeated}())) == Array{Int64,1}
+    @test eltype(permutations(Int[], 0, Val{:repeated}())) == Array{Int,1}
 
-    @test collect(permutations(Int[], 1, Val{:repeated}())) == Array{Int64,1}[]
+    @test collect(permutations(Int[], 1, Val{:repeated}())) == Array{Int,1}[]
     @test length(permutations(Int[], 1, Val{:repeated}())) == 0
-    @test eltype(permutations(Int[], 1, Val{:repeated}())) == Array{Int64,1}
+    @test eltype(permutations(Int[], 1, Val{:repeated}())) == Array{Int,1}
 
     @test collect(permutations(["a", "b", "c", "d"], 2, Val{:repeated}())) == Array[
         ["a", "a"],
